delete input-select & create select component
