+++ packages/ui/src/components/select/select-trigger.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import { Select as SelectPrimitive } from 'bits-ui'
 
-  import { cn } from '../../../utils'
+  import { cn } from '../../utils'
 
   type $$Props = SelectPrimitive.TriggerProps
   type $$Events = SelectPrimitive.TriggerEvents
